+++ packages/experimental/src/components/ContextHeader/ContextHeader.js
@@ -8,34 +8,40 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { Canary } from '../_Canary';
 import { pkg } from '../../settings';
 const componentName = 'ContextHeader';
 
-export const ContextHeader = !pkg.isComponentEnabled(componentName)
-  ? // Return canary if not released or flag not set
-    () => <Canary component={componentName} />
-  : // Main component code...
-    ({ namespace, name, task }) => {
-      return (
-        <div className={`${pkg.prefix}-context-header`}>
-          <span className={`${pkg.prefix}-context-header--task`}>{task}</span>
-          <span className={`${pkg.prefix}-context-header--name`}>{name}</span>
-          <span className={`${pkg.prefix}-context-header--separator`}>|</span>
-          <span className={`${pkg.prefix}-context-header--namespace`}>
-            {namespace}
-          </span>
-        </div>
-      );
-    };
+export let ContextHeader = ({ namespace, name, task }) => {
+  return (
+    <div className={`${pkg.prefix}-context-header`}>
+      <span className={`${pkg.prefix}-context-header--task`}>{task}</span>
+      <span className={`${pkg.prefix}-context-header--name`}>{name}</span>
+      <span className={`${pkg.prefix}-context-header--separator`}>|</span>
+      <span className={`${pkg.prefix}-context-header--namespace`}>
+        {namespace}
+      </span>
+    </div>
+  );
+};
 
-ContextHeader.displayName = 'ContextHeader';
+// Return a placeholder if not released and not enabled by feature flag
+ContextHeader = pkg.checkComponentEnabled(ContextHeader, componentName);
+
+ContextHeader.displayName = componentName;
 
 ContextHeader.propTypes = {
+  /**
+   * The name of the current instance.
+   */
   name: PropTypes.string.isRequired,
+  /**
+   * The namespace in which the current instance is deployed.
+   */
   namespace: PropTypes.string.isRequired,
+  /**
+   * The high level task the user is using the embedding UI to carry out.
+   */
   task: PropTypes.string.isRequired,
 };
-ContextHeader.displayName = componentName;
 
 export default ContextHeader;