+++ packages/experimental/src/global/js/utils/__tests__/unwrap-if-fragment.test.js
@@ -1,26 +1,39 @@
 import React from 'react';
 import unwrapIfFragment from '../unwrap-if-fragment.js';
 
+const AChild = () => <div>A child</div>;
+
 describe('unwrap-if-fragment', () => {
+  it('Should handle a fragment with one child not in an array', () => {
+    const result = unwrapIfFragment(
+      // The following disable is for test purposes
+      // eslint-disable-next-line react/jsx-no-useless-fragment
+      <>
+        <AChild />
+      </>
+    );
+
+    // The following disable is for test purposes
+    // eslint-disable-next-line react/jsx-key
+    expect(result).toEqual([<AChild />]);
+  });
+
   it('Should handle a fragment with one child', () => {
-    const child = 'a child';
-    const children = [child];
+    const children = [AChild];
     const result = unwrapIfFragment(<>{children.map((item) => item)}</>);
 
     expect(result).toEqual(children);
   });
 
   it('Should handle a fragment with multiple children', () => {
-    const child = 'a child';
-    const children = [child, child, child];
+    const children = [AChild, AChild, AChild];
     const result = unwrapIfFragment(<>{children.map((item) => item)}</>);
 
     expect(result).toEqual(children);
   });
 
   it('Should handle a multiple fragments and children', () => {
-    const child = 'a child';
-    const children = [child, child, child];
+    const children = [AChild, AChild, AChild];
     children.push([...children]);
     const result = unwrapIfFragment(<>{children.map((item) => item)}</>);
 
@@ -28,47 +41,42 @@ describe('unwrap-if-fragment', () => {
   });
 
   it('Should handle a nested fragments and children', () => {
-    const child = 'a child';
-    const children = [child, child, child];
+    const children = [AChild, AChild, AChild];
     children.push([...children]);
     const result = unwrapIfFragment(
       <>
         <>{children.map((item) => item)}</>
-        {child}
+        {AChild}
       </>
     );
 
     const expectedResult = children.flatMap((child) => child);
-    expectedResult.push(child);
+    expectedResult.push(AChild);
     expect(result).toEqual(expectedResult);
   });
 
-  it('Should handle an array with a lone one child', () => {
-    const child = 'a child';
-    const result = unwrapIfFragment(child);
+  it('Should handle a lone child', () => {
+    const result = unwrapIfFragment(AChild);
 
-    expect(result).toEqual([child]);
+    expect(result).toEqual([AChild]);
   });
 
   it('Should handle an array with one child', () => {
-    const child = 'a child';
-    const children = [child];
-    const result = unwrapIfFragment(child);
+    const children = [AChild];
+    const result = unwrapIfFragment(AChild);
 
     expect(result).toEqual(children);
   });
 
   it('Should handle an array with multiple children', () => {
-    const child = 'a child';
-    const children = [child, child, child];
+    const children = [AChild, AChild, AChild];
     const result = unwrapIfFragment(children);
 
     expect(result).toEqual(children);
   });
 
   it('Should handle an array with multiple children and levels', () => {
-    const child = 'a child';
-    const children = [child, child, child];
+    const children = [AChild, AChild, AChild];
     children.push(...children);
     const result = unwrapIfFragment(children);